+++ src/main/java/org/oagi/srt/gateway/http/api/cc_management/data/CcCreateResponse.java
@@ -2,9 +2,11 @@ package org.oagi.srt.gateway.http.api.cc_management.data;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class CcCreateResponse {
 
-    private long manifestId;
+    private BigInteger manifestId;
 
 }
